News App Working commit
