--- NewsWithWeather/app/src/main/java/com/example/newswithweather/CategoryClickListener.kt
@@ -1,5 +1,0 @@
-package com.example.newswithweather
-
-interface CategoryClickListener {
-    fun onCategoryClicked(category: String)
-}
