--- NewsWithWeather/app/src/main/java/com/example/newswithweather/SplashActivity.kt
@@ -1,27 +1,0 @@
-package com.example.newswithweather
-
-import android.content.Intent
-import android.os.Bundle
-import android.os.Handler
-import android.widget.ImageView
-import androidx.activity.enableEdgeToEdge
-import androidx.appcompat.app.AppCompatActivity
-import androidx.core.view.ViewCompat
-import androidx.core.view.WindowInsetsCompat
-import com.bumptech.glide.Glide
-
-class SplashActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setTheme(R.style.LightWithoutAppBar)
-        setContentView(R.layout.activity_splash)
-        val imageView : ImageView = findViewById(R.id.animationView)
-        Glide.with(this).load(R.drawable.splash_loading).into(imageView)
-        Handler().postDelayed({
-            val intent = Intent(this, MainActivity::class.java)
-            startActivity(intent)
-            finish()
-        }, 5000
-        )
-    }
-}
