+++ NewsWithWeather/app/src/main/java/com/example/newswithweather/activity/MainActivity.kt
@@ -1,0 +1,12 @@
+package com.example.newswithweather.activity
+
+import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+import com.example.newswithweather.R
+
+class MainActivity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_main)
+    }
+}
