+++ NewsWithWeather/app/src/main/java/com/example/newswithweather/listener/CategoryClickListener.kt
@@ -1,0 +1,5 @@
+package com.example.newswithweather.listener
+
+interface CategoryClickListener {
+    fun onCategoryClicked(category: String)
+}
