+++ NewsWithWeather/app/src/main/java/com/example/newswithweather/listener/ImageClickListener.kt
@@ -1,0 +1,7 @@
+package com.example.newswithweather.listener
+
+import android.widget.ImageView
+
+interface ImageClickListener {
+    fun onClick(imageUrl: String)
+}
