+++ NewsWithWeather/app/src/main/res/layout-land/fragment_list_view.xml
@@ -6,18 +6,17 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".ListViewFragment">
+        tools:context=".fragment.ListViewFragment">
 
         <androidx.recyclerview.widget.RecyclerView
             android:id="@+id/horizontalRecyclerView"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:layout_marginTop="16dp"
             android:orientation="horizontal"
             app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
-            app:layout_constraintTop_toTopOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toTopOf="parent"
             tools:listitem="@layout/category_recycler" />
 
         <androidx.recyclerview.widget.RecyclerView
@@ -41,7 +40,7 @@
             android:layout_marginEnd="84dp"
             android:layout_marginBottom="4dp"
             android:elevation="4dp"
-            android:text="29 C"
+            android:text="@string/_25_c"
             app:icon="@drawable/cloudy_2932604"
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintEnd_toEndOf="parent" />
