+++ NewsWithWeather/app/src/main/res/layout/category_recycler.xml
@@ -7,7 +7,7 @@
     android:layout_height="wrap_content"
     android:layout_marginHorizontal="16dp"
     android:layout_marginVertical="16dp"
-    app:cardBackgroundColor="#7497C2"
+    app:cardBackgroundColor="#FFF9F9"
     app:cardCornerRadius="10dp"
     app:cardElevation="16dp">
 
@@ -19,7 +19,7 @@
         android:layout_marginVertical="16dp"
         android:text="All"
         android:textAlignment="center"
-        android:textColor="#F4E2E2"
+        android:textColor="#7497C2"
         android:textSize="16sp"
         android:textStyle="bold" />
