+++ NewsWithWeather/app/src/main/res/layout/fragment_list_view.xml
@@ -6,11 +6,11 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        tools:context=".ListViewFragment">
+        tools:context=".fragment.ListViewFragment">
 
         <androidx.recyclerview.widget.RecyclerView
             android:id="@+id/horizontalRecyclerView"
-            android:layout_width="wrap_content"
+            android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:orientation="horizontal"
             app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
@@ -25,6 +25,9 @@
             android:id="@+id/verticalRecyclerView"
             android:layout_width="match_parent"
             android:layout_height="646dp"
+            android:layout_marginStart="8dp"
+            android:layout_marginTop="8dp"
+            android:layout_marginEnd="8dp"
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
@@ -36,10 +39,12 @@
             android:layout_width="410dp"
             android:layout_height="11dp"
             android:indeterminate="true"
-            android:visibility="gone"
+            android:visibility="invisible"
+            app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
-            tools:layout_editor_absoluteY="86dp" />
+            app:layout_constraintTop_toTopOf="parent"
+            app:layout_constraintVertical_bias="0.110000014" />
 
         <com.google.android.material.floatingactionbutton.ExtendedFloatingActionButton
             android:id="@+id/fab"
@@ -48,7 +53,7 @@
             android:layout_gravity="bottom|end"
             android:layout_margin="16dp"
             android:elevation="4dp"
-            android:text="29 C"
+            android:text="@string/_25_c"
             app:icon="@drawable/cloudy_2932604"
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
