+++ NewsWithWeather/app/src/main/res/layout/fragment_web_view.xml
@@ -5,7 +5,7 @@
 <androidx.constraintlayout.widget.ConstraintLayout
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".WebView">
+    tools:context=".fragment.WebView">
 
 
     <WebView
