+++ NewsWithWeather/app/src/main/res/values/colors.xml
@@ -7,4 +7,5 @@
     <color name="light_blue_A200">#FF40C4FF</color>
     <color name="light_blue_A400">#FF00B0FF</color>
     <color name="black_overlay">#66000000</color>
+    <color name="transparent">#00000000</color>
 </resources>
