+++ NewsWithWeather/app/src/main/res/values/strings.xml
@@ -6,4 +6,7 @@
     <string name="search">Search</string>
     <string name="dummy_button">Dummy Button</string>
     <string name="dummy_content">DUMMY\nCONTENT</string>
+    <!-- TODO: Remove or change this placeholder text -->
+    <string name="hello_blank_fragment">Hello blank fragment</string>
+    <string name="_25_c">25 °C</string>
 </resources>
